[RemoveUnnecessoryCode] - RemoveCode and Search Geolocate
