+++ pubspec.yaml
@@ -34,8 +34,7 @@ dependencies:
   dio: ^4.0.6
   logger: ^1.1.0
   proj4dart: ^2.0.0
-  # UI
-  solid_bottom_sheet: ^0.1.10
+
   flutter_naver_map:
     git: https://github.com/note11g/flutter_naver_map.git
   flutter_native_splash: ^2.2.7 # splah(첫 로고 화면)용 library
